requirements.txt
alpaca-py==0.43.2
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
ta>=0.10.0
matplotlib>=3.8.0
plotly>=5.18.0
Flask>=3.0.0
requests>=2.31.0
joblib>=1.3.0
python-dotenv>=1.0.0
pytz>=2024.1
